chore: restructure themed native components
