+++ components/themed-native/index.ts
@@ -1,0 +1,4 @@
+export * from './use-theme-color';
+
+export { default as ThemedText } from './themed-text';
+export { default as ThemedView } from './themed-view';
