+++ components/themed-native/themed-view.tsx
@@ -1,0 +1,21 @@
+import { View, type ViewProps } from 'react-native';
+//
+import { useThemeColor } from './use-theme-color';
+
+// ----------------------------------------------------------------------
+
+type ThemedViewProps = ViewProps & {
+  lightColor?: string;
+  darkColor?: string;
+};
+
+export default function ThemedView({
+  style,
+  lightColor,
+  darkColor,
+  ...otherProps
+}: ThemedViewProps) {
+  const backgroundColor = useThemeColor({ light: lightColor, dark: darkColor }, 'background');
+
+  return <View style={[{ backgroundColor }, style]} {...otherProps} />;
+}
